requirements.txt
googlemaps
python-dotenv
python-google-places
langchain-google-genai
langchain
nltk
pandas
fastapi
uvicorn
gunicorn
pydantic